+++ release-notes.md
@@ -1,3 +1,6 @@
+`foundry-7_release-1` - ???
+* NEW FEATURE: Add ability to show minion group (max) size
+
 `0.1.11` - 2021-06-29
 * Fix for missing localizations ([#50](https://github.com/wrycu/StarWarsFFG-Enhancements/issues/50))
    * Thanks `@Belgarath` for translating to French and pointing out missing localizations